+++ src/Phoenix.Shared/Extensions/FloatExtensions.cs
@@ -4,7 +4,7 @@ namespace Phoenix.Shared.Extensions
 {
    public static class FloatExtensions
    {
-      public static float Round(this float value, int digits = 1)
+      public static float Round(this float value, byte digits = 1)
       {
          return MathF.Round(value, digits);
       }
